+++ e2e/cases/manifest/index.test.ts
@@ -1,6 +1,7 @@
 import { join } from 'path';
 import { expect, test } from '@playwright/test';
 import { build } from '@scripts/shared';
+import { pluginReact } from '@rsbuild/plugin-react';
 
 const fixtures = __dirname;
 
@@ -10,6 +11,7 @@ test('enableAssetManifest', async () => {
     entry: {
       main: join(fixtures, 'src/index.jsx'),
     },
+    plugins: [pluginReact()],
     rsbuildConfig: {
       output: {
         enableAssetManifest: true,
