--- e2e/cases/source/global-vars/src/App.tsx
@@ -1,10 +1,0 @@
-if (ENABLE_TEST === true) {
-  const test = document.createElement('div');
-  test.id = 'test-el';
-  test.innerHTML = 'aaaaa';
-  document.body.appendChild(test);
-}
-
-const App = () => <div id="test">Hello Rsbuild!</div>;
-
-export default App;
