--- e2e/cases/source/resolve-extension-prefix/src/App.jsx
@@ -1,5 +1,0 @@
-function App() {
-  return <div id="test">Hello Rsbuild!</div>;
-}
-
-export default App;
