+++ e2e/cases/source/resolve-extension-prefix/src/index.js
@@ -1,7 +1,7 @@
-/* eslint-disable no-undef */
-import React from 'react';
-import { render } from 'react-dom';
-import App from './App';
 import './ex';
 
-render(React.createElement(App), document.getElementById('root'));
+const testEl = document.createElement('div');
+testEl.id = 'test';
+testEl.innerHTML = `Hello Rsbuild!`;
+
+document.body.appendChild(testEl);
