--- e2e/cases/source/tsconfig-paths/src/App.tsx
@@ -1,10 +1,0 @@
-import { content } from '@/common/test';
-
-const testAliasEl = document.createElement('div');
-testAliasEl.id = 'foo';
-testAliasEl.innerHTML = content;
-document.body.appendChild(testAliasEl);
-
-const App = () => <div id="test">Hello Rsbuild!</div>;
-
-export default App;
