+++ e2e/cases/sri/index.test.ts
@@ -2,12 +2,14 @@ import path from 'path';
 import { expect } from '@playwright/test';
 import { build, getHrefByEntryName } from '@scripts/shared';
 import { webpackOnlyTest } from '@scripts/helper';
+import { pluginReact } from '@rsbuild/plugin-react';
 
 webpackOnlyTest('security.sri', async ({ page }) => {
   const rsbuild = await build({
     cwd: __dirname,
     entry: { index: path.resolve(__dirname, './src/index.js') },
     runServer: true,
+    plugins: [pluginReact()],
     rsbuildConfig: {
       security: {
         sri: true,
